--- a/pkg/controller/nodeipam/ipam/cidrset/BUILD
+++ b/pkg/controller/nodeipam/ipam/cidrset/BUILD
@@ -10,13 +10,23 @@ go_test(
     name = "go_default_test",
     srcs = ["cidr_set_test.go"],
     embed = [":go_default_library"],
-    deps = ["//vendor/k8s.io/klog/v2:go_default_library"],
+    deps = [
+        "//staging/src/k8s.io/component-base/metrics/testutil:go_default_library",
+        "//vendor/k8s.io/klog/v2:go_default_library",
+    ],
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["cidr_set.go"],
+    srcs = [
+        "cidr_set.go",
+        "metrics.go",
+    ],
     importpath = "k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset",
+    deps = [
+        "//staging/src/k8s.io/component-base/metrics:go_default_library",
+        "//staging/src/k8s.io/component-base/metrics/legacyregistry:go_default_library",
+    ],
 )
 
 filegroup(