<!--  Thanks for sending a pull request!  Here are some tips for you:

1. If this is your first time, please read our contributor guidelines: https://git.k8s.io/community/contributors/guide/first-contribution.md#your-first-contribution and developer guide https://git.k8s.io/community/contributors/devel/development.md#development-guide
2. Please label this pull request according to what type of issue you are addressing, especially if this is a release targeted pull request. For reference on required PR/issue labels, read here:
https://git.k8s.io/community/contributors/devel/sig-release/release.md#issuepr-kind-label
3. Ensure you have added or ran the appropriate tests for your PR: https://git.k8s.io/community/contributors/devel/sig-testing/testing.md
4. If you want *faster* PR reviews, read how: https://git.k8s.io/community/contributors/guide/pull-requests.md#best-practices-for-faster-reviews
5. If the PR is unfinished, see how to mark it: https://git.k8s.io/community/contributors/guide/pull-requests.md#marking-unfinished-pull-requests
-->

**What type of PR is this?**
> Uncomment only one ` /kind <>` line, hit enter to put that in a new line, and remove leading whitespace from that line:
>
> /kind api-change

/kind bug
/kind cleanup

> /kind deprecation
> /kind design
> /kind documentation
> /kind failing-test
> /kind feature
> /kind flake

**What this PR does / why we need it**:

    Improve range allocator cidr_set performance
    
    The range allocator that was using cidr_sets had a performance
    problem because, when the bitmap was full, it had to iterate over
    the whole bitmap to check that.
    
    Adding a counter to the cidr_set struct, we can track the number
    of allocated ranges and, in case it is full, return fast.
    
    The benchmarks added show the improvement:
    
```
$ /usr/local/go/bin/go test -benchmem -run=^$ k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset -bench . -cpuprofile
=cpu.out -v
    goos: darwin
    goarch: amd64
    pkg: k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset
    BenchmarkAllocateNext_48_52-4           27519943                39.8 ns/op             0 B/op          0 allocs/op
    BenchmarkAllocateNext_48_56-4           29308615                42.0 ns/op             0 B/op          0 allocs/op
    BenchmarkAllocateNext_48_60-4           28926524                39.9 ns/op             0 B/op          0 allocs/op
    BenchmarkAllocateNext_48_64-4           29793016                39.9 ns/op             0 B/op          0 allocs/op
    PASS
    ok      k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset  5.864s
```

Previously, the benchmark showed the dependency on the size of the benchmark:

```
$ /usr/local/go/bin/go test -benchmem -run=^$ k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset -bench .
    goos: darwin
    goarch: amd64
    pkg: k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset
    BenchmarkAllocateNext_48_52-4            2190278               516 ns/op               0 B/op          0 allocs/op
    BenchmarkAllocateNext_48_56-4             216110              6212 ns/op               0 B/op          0 allocs/op
    BenchmarkAllocateNext_48_60-4              20401             76832 ns/op              17 B/op          0 allocs/op
    PASS
    ok      k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset  5.279s
    
    The BenchmarkAllocateNext_48_64-4 is disabled because it takes more than
    60s
```

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

**Special notes for your reviewer**:

This shouldn't be a problem if the range is never fully allocated

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required:
Enter your extended release note in the block below. If the PR requires additional action from users switching to the new release, include the string "action required".

For more information on release notes see: https://git.k8s.io/community/contributors/guide/release-notes.md
-->
```release-note
improve the range allocator, based in cidr_sets, performance
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:

<!--
This section can be blank if this pull request does not require a release note.

When adding links which point to resources within git repositories, like
KEPs or supporting documentation, please reference a specific commit and avoid
linking directly to the master branch. This ensures that links reference a
specific point in time, rather than a document that may change over time.

See here for guidance on getting permanent links to files: https://help.github.com/en/articles/getting-permanent-links-to-files

Please use the following format for linking documentation:
- [KEP]: <link>
- [Usage]: <link>
- [Other doc]: <link>
-->
```docs

```