@aojea 2020-04-19
/assign @wojtek-t @thockin 

I've found this issue by chance, it shouldn't be a problem if the range is not fully allocated, but I let you decide if is worth to fix it.