@aojea 2020-04-21
New benchmark allocating all + 1 and then  releasing all

```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset
BenchmarkAllocateNext_48_52-8   	  156325	      7291 ns/op	    1424 B/op	      65 allocs/op
BenchmarkAllocateNext_48_56-8   	   13998	    108981 ns/op	   22544 B/op	    1025 allocs/op
BenchmarkAllocateNext_48_60-8   	     633	   1859388 ns/op	  360471 B/op	   16385 allocs/op
BenchmarkAllocateNext_48_64-8   	      56	  18831169 ns/op	 5783310 B/op	  262148 allocs/op
```