@wojtek-t 2020-04-21
Please update PR description with the results of the new benchmark both with and without this PR.