@aojea 2020-04-21
> Please update PR description with the results of the new benchmark both with and without this PR.

same number of allocations, and a little bit of improvement

BEFORE 
```sh
/usr/local/go/bin/go test -benchmem -run=^$ k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset -bench . -v -benchtime 30s 
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset
BenchmarkAllocateNext_48_52-8            6923869              5661 ns/op            1424 B/op         65 allocs/op
BenchmarkAllocateNext_48_56-8             350475             85874 ns/op           22544 B/op       1025 allocs/op
BenchmarkAllocateNext_48_60-8              25971           1380570 ns/op          360465 B/op      16385 allocs/op
BenchmarkAllocateNext_48_64-8               1778          20787282 ns/op         5767719 B/op     262145 allocs/op
PASS
ok      k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset  164.509s

AFTER
```sh
/usr/local/go/bin/go test -benchmem -run=^$ k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset -bench . -v -benchtime 30s 
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset
BenchmarkAllocateNext_48_52-8            6028858              5079 ns/op            1424 B/op         65 allocs/op
BenchmarkAllocateNext_48_56-8             451299             75592 ns/op           22544 B/op       1025 allocs/op
BenchmarkAllocateNext_48_60-8              29452           1289312 ns/op          360465 B/op      16385 allocs/op
BenchmarkAllocateNext_48_64-8               1590          22773748 ns/op         5767781 B/op     262145 allocs/op
PASS
ok      k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset  160.448s
```