@aojea 2020-04-21
/remove-kind bug