@bowei 2020-04-24
Can you update the benchmark results in the description and label the columns. I'm assuming the B/s is bytes/S? is this incremental increase or total.