@aojea 2020-04-26
> Can you update the benchmark results in the description and label the columns. I'm assuming the B/s is bytes/S? is this incremental increase or total.

The benchmarks are obtained with `/usr/local/go/bin/go test k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset -benchmem -bench .`

The fields meaning are explained here https://stackoverflow.com/a/35588683/7794348
```
goos: linux
goarch: amd64
pkg: k8s.io/kubernetes/pkg/controller/nodeipam/ipam/cidrset
BenchmarkAllocateAll_48_52-8       76086             15501 ns/op            2080 B/op         98 allocs/op
BenchmarkAllocateAll_48_56-8        5023            240764 ns/op           32800 B/op       1538 allocs/op
BenchmarkAllocateAll_48_60-8         308           3896779 ns/op          524336 B/op      24578 allocs/op
BenchmarkAllocateAll_48_64-8          19          62209311 ns/op         8436147 B/op     393229 allocs/op
BenchmarkAllocateAll_64_68-8       75800             15609 ns/op            2080 B/op         98 allocs/op
BenchmarkAllocateAll_64_72-8        4953            244595 ns/op           32800 B/op       1538 allocs/op
BenchmarkAllocateAll_64_76-8         304           3901199 ns/op          524337 B/op      24578 allocs/op
BenchmarkAllocateAll_64_80-8          19          62812786 ns/op         8436142 B/op     393229 allocs/op
PASS
ok  
```

To obtain the delta, I use https://godoc.org/golang.org/x/tools/cmd/benchcmp , just store the benchmark against the different version of cidr_set.go

Seems that adding the prometheus metrics has a big impact :/
```
benchmark                        old ns/op     new ns/op     delta
BenchmarkAllocateAll_48_52-8     4035          29551         +632.37%
BenchmarkAllocateAll_48_56-8     60501         458525        +657.88%
BenchmarkAllocateAll_48_60-8     956916        7359604       +669.10%
BenchmarkAllocateAll_48_64-8     15253670      118401352     +676.22%
BenchmarkAllocateAll_64_68-8     3982          29404         +638.42%
BenchmarkAllocateAll_64_72-8     59333         461355        +677.57%
BenchmarkAllocateAll_64_76-8     944095        7372590       +680.92%
BenchmarkAllocateAll_64_80-8     15031076      118523465     +688.52%

benchmark                        old allocs     new allocs     delta
BenchmarkAllocateAll_48_52-8     65             293            +350.77%
BenchmarkAllocateAll_48_56-8     1025           4613           +350.05%
BenchmarkAllocateAll_48_60-8     16385          73733          +350.00%
BenchmarkAllocateAll_48_64-8     262147         1179677        +350.01%
BenchmarkAllocateAll_64_68-8     65             293            +350.77%
BenchmarkAllocateAll_64_72-8     1025           4613           +350.05%
BenchmarkAllocateAll_64_76-8     16385          73733          +350.00%
BenchmarkAllocateAll_64_80-8     262147         1179677        +350.01%

benchmark                        old bytes     new bytes     delta
BenchmarkAllocateAll_48_52-8     1424          4160          +192.13%
BenchmarkAllocateAll_48_56-8     22544         65600         +190.99%
BenchmarkAllocateAll_48_60-8     360469        1048671       +190.92%
BenchmarkAllocateAll_48_64-8     5779232       16877760      +192.04%
BenchmarkAllocateAll_64_68-8     1424          4160          +192.13%
BenchmarkAllocateAll_64_72-8     22544         65600         +190.99%
BenchmarkAllocateAll_64_76-8     360469        1048670       +190.92%
BenchmarkAllocateAll_64_80-8     5778772       16877514      +192.06%
```

if we compare about the PR without the new metrics

```
benchcmp  old mid.txt
benchmark                        old ns/op     new ns/op     delta
BenchmarkAllocateAll_48_52-8     4035          2529          -37.32%
BenchmarkAllocateAll_48_56-8     60501         36971         -38.89%
BenchmarkAllocateAll_48_60-8     956916        584870        -38.88%
BenchmarkAllocateAll_48_64-8     15253670      9384728       -38.48%
BenchmarkAllocateAll_64_68-8     3982          2578          -35.26%
BenchmarkAllocateAll_64_72-8     59333         38231         -35.57%
BenchmarkAllocateAll_64_76-8     944095        605037        -35.91%
BenchmarkAllocateAll_64_80-8     15031076      9802871       -34.78%

benchmark                        old allocs     new allocs     delta
BenchmarkAllocateAll_48_52-8     65             33             -49.23%
BenchmarkAllocateAll_48_56-8     1025           513            -49.95%
BenchmarkAllocateAll_48_60-8     16385          8193           -50.00%
BenchmarkAllocateAll_48_64-8     262147         131074         -50.00%
BenchmarkAllocateAll_64_68-8     65             33             -49.23%
BenchmarkAllocateAll_64_72-8     1025           513            -49.95%
BenchmarkAllocateAll_64_76-8     16385          8193           -50.00%
BenchmarkAllocateAll_64_80-8     262147         131074         -50.00%

benchmark                        old bytes     new bytes     delta
BenchmarkAllocateAll_48_52-8     1424          1040          -26.97%
BenchmarkAllocateAll_48_56-8     22544         16400         -27.25%
BenchmarkAllocateAll_48_60-8     360469        262162        -27.27%
BenchmarkAllocateAll_48_64-8     5779232       4201381       -27.30%
BenchmarkAllocateAll_64_68-8     1424          1040          -26.97%
BenchmarkAllocateAll_64_72-8     22544         16400         -27.25%
BenchmarkAllocateAll_64_76-8     360469        262162        -27.27%
BenchmarkAllocateAll_64_80-8     5778772       4201725       -27.29%
```

maybe the improvement is relatively significant, but not practically meaningful? 🤷